// src/types/partials/ContractTerm.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TDateTime } from '../simpleTypes/DateTime';
import { TPeriod } from '../partials/Period';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TContractSecurityLabel } from '../partials/ContractSecurityLabel';
import { TContractOffer } from '../partials/ContractOffer';
import { TContractAsset } from '../partials/ContractAsset';
import { TContractAction } from '../partials/ContractAction';

export type TContractTerm = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    identifier?: TIdentifier;
    issued?: TDateTime;
    applies?: TPeriod;
    topicCodeableConcept?: TCodeableConcept;
    topicReference?: TReference;
    type?: TCodeableConcept;
    subType?: TCodeableConcept;
    text?: String;
    securityLabel?: TContractSecurityLabel[];
    offer: TContractOffer;
    asset?: TContractAsset[];
    action?: TContractAction[];
    group?: TContractTerm[];
};